requirements.txt
python-telegram-bot==13.7
pymongo[srv]==3.12.0
python-dotenv==0.19.0
requests==2.26.0

